Refactor: Pages 폴더 서비스와 관리자 분리
