+++ src/pages/Service/CommunityDetail/CommunityDetail.jsx
