+++ src/pages/Service/CommunityDetail/CommunityDetail.style.js
