+++ src/pages/Service/CommunityRegist/CommunityRegistPage.style.js
