+++ src/pages/Service/Lecture/constants.js
